+++ R/availableWorkers.R
@@ -215,10 +215,12 @@ availableWorkers <- function(constraints = NULL, methods = getOption2("parallell
       }
       w <- read_pe_hostfile(pathname, expand = TRUE)
 
-      ## Sanity checks
+      ## Sanity checks: It is not always true that length(w) == $NSLOTS, e.g.
+      ## on the UCSF Wynton SGE cluster, 'qsub -pe mpi-8 16 ...' will produce
+      ## a job with w=2 workers and NSLOTS=16. /HB 2023-02-01
       nslots <- as.integer(getenv("NSLOTS"))
-      if (!identical(nslots, length(w))) {
-        warnf("Identified %d workers from the %s file (%s), which does not match environment variable %s = %d", length(w), sQuote("PE_HOSTFILE"), sQuote(pathname), sQuote("NSLOTS"), nslots)
+      if (nslots >= length(w)) {
+        warnf("Identified %d workers from the %s file (%s), which is more than environment variable %s = %d", length(w), sQuote("PE_HOSTFILE"), sQuote(pathname), sQuote("NSLOTS"), nslots)
       }
     } else if (method == "Slurm") {
       ## From 'man sbatch':